refactor(BA-5982): drop deployment_revision adapter split
Roll back the dedicated ``DeploymentRevisionAdapter`` introduced earlier
in this PR. Adapter / handler / processor for the bulk admin refresh all
live back inside the single ``deployment`` slice; the service-side
helper ``_build_creator_from_revision_data`` and the inline
``admin_refresh_deployment_revisions`` loop are restored to their
pre-split form, and ``RefreshDeploymentRevisionsAction`` returns to the
no-arg shape that triggers the service to do the whole sweep on its
own. ``DeploymentAdapter`` is again the single owner of revision-side
adapter methods including ``admin_refresh_deployment_revisions``; the
GQL resolver and REST handler call it directly with no extra
``revision_adapter`` injection.

Other refactor work in this PR (``from_spec`` →
``from_revision_spec``, ``replica_counts`` → ``replica``, the
``ModelRevisionData.deployment_id`` field, the Spec/Data split, the
``revision_number`` exposure) is unaffected.
