+++ src/ai/backend/manager/services/deployment/service.py
@@ -20,14 +20,20 @@ from ai.backend.common.dto.appproxy_coordinator.v2.endpoint.request import (
 from ai.backend.common.identifier.deployment import DeploymentID
 from ai.backend.logging.utils import BraceStyleAdapter
 from ai.backend.manager.clients.appproxy.client import AppProxyClientPool
+from ai.backend.manager.data.deployment.creator import (
+    ModelRevisionCreator,
+    VFolderMountsCreator,
+)
 from ai.backend.manager.data.deployment.types import (
     DeploymentInfo,
+    ExecutionSpec,
     ModelDeploymentAccessTokenData,
     ModelDeploymentData,
     ModelDeploymentMetadataInfo,
     ModelReplicaData,
     ModelRevisionData,
     ReplicaStateData,
+    ResourceSpec,
     RevisionRefreshResult,
     RouteHealthStatus,
     RouteInfo,
@@ -41,6 +47,8 @@ from ai.backend.manager.models.deployment_policy import (
     DeploymentPolicyRow,
 )
 from ai.backend.manager.models.endpoint import EndpointTokenRow
+from ai.backend.manager.models.endpoint.conditions import DeploymentConditions
+from ai.backend.manager.repositories.base import BatchQuerier, NoPagination
 from ai.backend.manager.repositories.base.rbac.entity_creator import RBACEntityCreator
 from ai.backend.manager.repositories.base.upserter import Upserter
 from ai.backend.manager.repositories.deployment import DeploymentRepository
@@ -316,6 +324,51 @@ def _convert_route_info_to_replica_data(route: RouteInfo) -> ModelReplicaData:
         activeness_status=_resolve_activeness(route.traffic_status, readiness, liveness),
         detail=route.error_data,
         created_at=route.created_at,
+    )
+
+
+def _build_creator_from_revision_data(data: ModelRevisionData) -> ModelRevisionCreator:
+    """Rebuild a ``ModelRevisionCreator`` from an existing revision's persisted data.
+
+    ``model_definition`` is reset to ``None`` so ``DeploymentController.add_revision``
+    re-resolves it from the vfolder. ``revision_preset_id`` and ``preset_values`` are
+    carried over so the new revision keeps the same preset attribution. ``extra_mounts``
+    is left empty because ``add_revision`` does not propagate this field to the new
+    revision spec.
+    """
+    if data.model_mount_config.vfolder_id is None:
+        raise InvalidAPIParameters(
+            f"Revision {data.id} has no model vfolder; cannot rebuild creator"
+        )
+    return ModelRevisionCreator(
+        image_id=data.image_id,
+        resource_spec=ResourceSpec(
+            cluster_mode=data.cluster_config.mode,
+            cluster_size=data.cluster_config.size,
+            resource_slots=dict(data.resource_config.resource_slot),
+            resource_opts=dict(data.resource_config.resource_opts) or None,
+        ),
+        mounts=VFolderMountsCreator(
+            model_vfolder_id=data.model_mount_config.vfolder_id,
+            model_definition_path=data.model_mount_config.definition_path or None,
+            model_mount_destination=data.model_mount_config.mount_destination or "/models",
+            extra_mounts=[],
+        ),
+        execution=ExecutionSpec(
+            startup_command=data.startup_command,
+            bootstrap_script=data.bootstrap_script,
+            environ=(
+                {k: str(v) for k, v in data.model_runtime_config.environ.items()}
+                if data.model_runtime_config.environ
+                else None
+            ),
+            runtime_variant_id=data.model_runtime_config.runtime_variant_id,
+            callback_url=data.callback_url,
+            inference_runtime_config=data.model_runtime_config.inference_runtime_config,
+        ),
+        model_definition=None,
+        revision_preset_id=data.revision_preset_id,
+        preset_values=list(data.preset_values),
     )
 
 
@@ -606,34 +659,54 @@ class DeploymentService:
             deployment_policy=result.deployment_policy,
         )
 
-    async def refresh_deployment_revisions(
-        self,
-        action: RefreshDeploymentRevisionsAction,
+    async def admin_refresh_deployment_revisions(
+        self, action: RefreshDeploymentRevisionsAction
     ) -> RefreshDeploymentRevisionsActionResult:
-        """Add and activate a fresh revision for each entry in
-        ``action.creators_by_id``.
+        """Refresh revisions for all active deployments.
 
-        The adapter has already converted every previous-revision
-        ``ModelRevisionData`` into a ``ModelRevisionCreator``; this method
-        only runs ``DeploymentController.add_deployment_revision(
-        auto_activate=True)`` per (deployment_id, creator) pair and packages
-        the outcomes into ``RevisionRefreshResult``s. Each deployment is
-        processed independently so a single failure does not abort the rest
-        (partial success by design).
+        For each active deployment, rebuilds a ``ModelRevisionCreator`` from the
+        current revision and delegates to
+        ``DeploymentController.add_deployment_revision(auto_activate=True)``
+        (which re-resolves preset / deployment-config / model_definition and
+        activates the new revision). Each deployment is processed independently
+        so a single failure does not abort the rest (partial success by design).
         """
+        # Bulk scan + independent per-deployment orchestration: multiple repo
+        # and controller calls are required by design to preserve partial
+        # success semantics. Each inner call owns its own transaction boundary.
+        active_querier = BatchQuerier(
+            pagination=NoPagination(),
+            conditions=[
+                DeploymentConditions.by_lifecycle_stages(EndpointLifecycle.active_states()),
+            ],
+        )
+        deployment_ids = await self._deployment_repository.search_deployment_ids(
+            querier=active_querier,
+        )
         results: list[RevisionRefreshResult] = []
         succeeded = 0
         failed = 0
-        for deployment_id, creator in action.creators_by_id.items():
+        for deployment_id in deployment_ids:
             try:
+                data = await self._deployment_repository.get_current_revision(deployment_id)
+                creator = _build_creator_from_revision_data(data)
                 new_revision = await self._deployment_controller.add_deployment_revision(
                     deployment_id=deployment_id,
                     revision=creator,
                     auto_activate=True,
                 )
+                results.append(
+                    RevisionRefreshResult(
+                        deployment_id=deployment_id,
+                        new_revision_id=new_revision.id,
+                        success=True,
+                        failure_reason=None,
+                    )
+                )
+                succeeded += 1
             except Exception as exc:
                 log.warning(
-                    "refresh_deployment_revisions failed for deployment {}: {}: {}",
+                    "admin_refresh_deployment_revisions failed for deployment {}: {}: {}",
                     deployment_id,
                     type(exc).__name__,
                     exc,
@@ -647,19 +720,9 @@ class DeploymentService:
                     )
                 )
                 failed += 1
-                continue
-            results.append(
-                RevisionRefreshResult(
-                    deployment_id=deployment_id,
-                    new_revision_id=new_revision.id,
-                    success=True,
-                    failure_reason=None,
-                )
-            )
-            succeeded += 1
         log.info(
-            "refresh_deployment_revisions summary: total={} succeeded={} failed={}",
-            len(action.creators_by_id),
+            "admin_refresh_deployment_revisions summary: total={} succeeded={} failed={}",
+            len(deployment_ids),
             succeeded,
             failed,
         )
